Split the API controller into 2
