+++ appinfo/info.xml
@@ -55,7 +55,7 @@ And in the works for the [coming versions](https://github.com/nextcloud/spreed/m
 	<navigations>
 		<navigation>
 			<name>Video calls</name>
-			<route>spreed.page.index</route>
+			<route>spreed.Page.index</route>
 			<order>3</order>
 		</navigation>
 	</navigations>
