+++ lib/Controller/CallController.php
@@ -28,7 +28,6 @@ namespace OCA\Spreed\Controller;
 use OCA\Spreed\Exceptions\RoomNotFoundException;
 use OCA\Spreed\Manager;
 use OCA\Spreed\Room;
-use OCP\Activity\IManager as IActivityManager;
 use OCP\AppFramework\Http;
 use OCP\AppFramework\Http\DataResponse;
 use OCP\AppFramework\OCSController;
@@ -38,29 +37,20 @@ use OCP\IRequest;
 use OCP\ISession;
 use OCP\IUser;
 use OCP\IUserManager;
-use OCP\IGroup;
-use OCP\IGroupManager;
-use OCP\Notification\IManager as INotificationManager;
 
-class ApiController extends OCSController {
+class CallController extends OCSController {
 	/** @var string */
 	private $userId;
 	/** @var IL10N */
 	private $l10n;
 	/** @var IUserManager */
 	private $userManager;
-	/** @var IGroupManager */
-	private $groupManager;
 	/** @var ISession */
 	private $session;
 	/** @var ILogger */
 	private $logger;
 	/** @var Manager */
 	private $manager;
-	/** @var INotificationManager */
-	private $notificationManager;
-	/** @var IActivityManager */
-	private $activityManager;
 
 	/**
 	 * @param string $appName
@@ -68,34 +58,25 @@ class ApiController extends OCSController {
 	 * @param IRequest $request
 	 * @param IL10N $l10n
 	 * @param IUserManager $userManager
-	 * @param IGroupManager $groupManager
 	 * @param ISession $session
 	 * @param ILogger $logger
 	 * @param Manager $manager
-	 * @param INotificationManager $notificationManager
-	 * @param IActivityManager $activityManager
 	 */
 	public function __construct($appName,
 								$UserId,
 								IRequest $request,
 								IL10N $l10n,
 								IUserManager $userManager,
-								IGroupManager $groupManager,
 								ISession $session,
 								ILogger $logger,
-								Manager $manager,
-								INotificationManager $notificationManager,
-								IActivityManager $activityManager) {
+								Manager $manager) {
 		parent::__construct($appName, $request);
 		$this->userId = $UserId;
 		$this->l10n = $l10n;
 		$this->userManager = $userManager;
-		$this->groupManager = $groupManager;
 		$this->session = $session;
 		$this->logger = $logger;
 		$this->manager = $manager;
-		$this->notificationManager = $notificationManager;
-		$this->activityManager = $activityManager;
 	}
 
 	/**
@@ -293,215 +274,6 @@ class ApiController extends OCSController {
 	}
 
 	/**
-	 * Initiates a one-to-one video call from the current user to the recipient
-	 *
-	 * @NoAdminRequired
-	 *
-	 * @param string $targetUserName
-	 * @return DataResponse
-	 */
-	public function createOneToOneRoom($targetUserName) {
-		// Get the user
-		$targetUser = $this->userManager->get($targetUserName);
-		$currentUser = $this->userManager->get($this->userId);
-		if(!($targetUser instanceof IUser)) {
-			return new DataResponse([], Http::STATUS_NOT_FOUND);
-		}
-
-		// If room exists: Reuse that one, otherwise create a new one.
-		try {
-			$room = $this->manager->getOne2OneRoom($this->userId, $targetUser->getUID());
-			return new DataResponse(['token' => $room->getToken()], Http::STATUS_OK);
-		} catch (RoomNotFoundException $e) {
-			$room = $this->manager->createOne2OneRoom();
-			$room->addUser($currentUser);
-
-			$room->addUser($targetUser);
-			$this->createNotification($currentUser, $targetUser, $room);
-
-			return new DataResponse(['token' => $room->getToken()], Http::STATUS_CREATED);
-		}
-	}
-
-	/**
-	 * Initiates a group video call from the selected group
-	 *
-	 * @NoAdminRequired
-	 *
-	 * @param string $targetGroupName
-	 * @return DataResponse
-	 */
-	public function createGroupRoom($targetGroupName) {
-		$targetGroup = $this->groupManager->get($targetGroupName);
-		$currentUser = $this->userManager->get($this->userId);
-
-		if(!($targetGroup instanceof IGroup)) {
-			return new DataResponse([], Http::STATUS_NOT_FOUND);
-		}
-
-		$usersInGroup = $targetGroup->getUsers();
-		// If the user who is creating this call is not part of this group add them
-		if (!$targetGroup->inGroup($currentUser)) {
-			$usersInGroup[] = $currentUser;
-		}
-
-		// Create the room
-		$room = $this->manager->createGroupRoom($targetGroup->getGID());
-		foreach ($usersInGroup as $user) {
-			$room->addUser($user);
-
-			if ($currentUser->getUID() !== $user->getUID()) {
-				$this->createNotification($currentUser, $user, $room);
-			}
-		}
-
-		return new DataResponse(['token' => $room->getToken()], Http::STATUS_CREATED);
-	}
-
-	/**
-	 * @NoAdminRequired
-	 *
-	 * @return DataResponse
-	 */
-	public function createPublicRoom() {
-		$currentUser = $this->userManager->get($this->userId);
-
-		// Create the room
-		$room = $this->manager->createPublicRoom();
-		$room->addUser($currentUser);
-
-		return new DataResponse(['token' => $room->getToken()], Http::STATUS_CREATED);
-	}
-
-	/**
-	 * @NoAdminRequired
-	 *
-	 * @param int $roomId
-	 * @param string $roomName
-	 * @return DataResponse
-	 */
-	public function renameRoom($roomId, $roomName) {
-		try {
-			$room = $this->manager->getRoomForParticipant($roomId, $this->userId);
-		} catch (RoomNotFoundException $e) {
-			return new DataResponse([], Http::STATUS_NOT_FOUND);
-		}
-
-		if (strlen($roomName) > 200) {
-			return new DataResponse([], Http::STATUS_BAD_REQUEST);
-		}
-
-		if (!$room->setName($roomName)) {
-			return new DataResponse([], Http::STATUS_METHOD_NOT_ALLOWED);
-		}
-		return new DataResponse([]);
-	}
-
-	/**
-	 * @NoAdminRequired
-	 *
-	 * @param int $roomId
-	 * @param string $newParticipant
-	 * @return DataResponse
-	 */
-	public function addParticipantToRoom($roomId, $newParticipant) {
-		try {
-			$room = $this->manager->getRoomForParticipant($roomId, $this->userId);
-		} catch (RoomNotFoundException $e) {
-			return new DataResponse([], Http::STATUS_NOT_FOUND);
-		}
-
-		$participants = $room->getParticipants();
-		if (isset($participants['users'][$newParticipant])) {
-			return new DataResponse([]);
-		}
-
-		$currentUser = $this->userManager->get($this->userId);
-		$newUser = $this->userManager->get($newParticipant);
-		if (!$newUser instanceof IUser) {
-			return new DataResponse([], Http::STATUS_NOT_FOUND);
-		}
-
-		if ($room->getType() === Room::ONE_TO_ONE_CALL) {
-			// In case a user is added to a one2one call, we change the call to a group call
-			$room->changeType(Room::GROUP_CALL);
-
-			$room->addUser($newUser);
-			$this->createNotification($currentUser, $newUser, $room);
-
-			return new DataResponse(['type' => $room->getType()]);
-		}
-
-		$room->addUser($newUser);
-		$this->createNotification($currentUser, $newUser, $room);
-
-		return new DataResponse([]);
-	}
-
-	/**
-	 * @NoAdminRequired
-	 *
-	 * @param int $roomId
-	 * @return DataResponse
-	 */
-	public function leaveRoom($roomId) {
-		try {
-			$room = $this->manager->getRoomForParticipant($roomId, $this->userId);
-		} catch (RoomNotFoundException $e) {
-			return new DataResponse([], Http::STATUS_NOT_FOUND);
-		}
-
-		if ($room->getType() === Room::ONE_TO_ONE_CALL || $room->getNumberOfParticipants() === 1) {
-			$room->deleteRoom();
-		} else {
-			$currentUser = $this->userManager->get($this->userId);
-			$room->removeUser($currentUser);
-		}
-
-		return new DataResponse([]);
-	}
-
-	/**
-	 * @NoAdminRequired
-	 *
-	 * @param int $roomId
-	 * @return DataResponse
-	 */
-	public function makePublic($roomId) {
-		try {
-			$room = $this->manager->getRoomForParticipant($roomId, $this->userId);
-		} catch (RoomNotFoundException $e) {
-			return new DataResponse([], Http::STATUS_NOT_FOUND);
-		}
-
-		if ($room->getType() !== Room::PUBLIC_CALL) {
-			$room->changeType(Room::PUBLIC_CALL);
-		}
-
-		return new DataResponse();
-	}
-
-	/**
-	 * @NoAdminRequired
-	 *
-	 * @param int $roomId
-	 * @return DataResponse
-	 */
-	public function makePrivate($roomId) {
-		try {
-			$room = $this->manager->getRoomForParticipant($roomId, $this->userId);
-		} catch (RoomNotFoundException $e) {
-			return new DataResponse([], Http::STATUS_NOT_FOUND);
-		}
-
-		if ($room->getType() === Room::PUBLIC_CALL) {
-			$room->changeType(Room::GROUP_CALL);
-		}
-
-		return new DataResponse();
-	}
-
-	/**
 	 * @PublicPage
 	 *
 	 * @param string $token
@@ -571,46 +343,4 @@ class ApiController extends OCSController {
 		return new DataResponse();
 	}
 
-	/**
-	 * @param IUser $actor
-	 * @param IUser $user
-	 * @param Room $room
-	 */
-	protected function createNotification(IUser $actor, IUser $user, Room $room) {
-		$notification = $this->notificationManager->createNotification();
-		$dateTime = new \DateTime();
-		try {
-			$notification->setApp('spreed')
-				->setUser($user->getUID())
-				->setDateTime($dateTime)
-				->setObject('room', $room->getId())
-				->setSubject('invitation', [$actor->getUID()]);
-			$this->notificationManager->notify($notification);
-		} catch (\InvalidArgumentException $e) {
-			// Error while creating the notification
-			$this->logger->logException($e, ['app' => 'spreed']);
-		}
-
-		$event = $this->activityManager->generateEvent();
-		try {
-			$event->setApp('spreed')
-				->setType('spreed')
-				->setAuthor($actor->getUID())
-				->setAffectedUser($user->getUID())
-				->setObject('room', $room->getId(), $room->getName())
-				->setTimestamp($dateTime->getTimestamp())
-				->setSubject('invitation', [
-					'user' => $actor->getUID(),
-					'room' => $room->getId(),
-					'name' => $room->getName(),
-				]);
-			$this->activityManager->publish($event);
-		} catch (\InvalidArgumentException $e) {
-			// Error while creating the activity
-			$this->logger->logException($e, ['app' => 'spreed']);
-		} catch (\BadMethodCallException $e) {
-			// Error while sending the activity
-			$this->logger->logException($e, ['app' => 'spreed']);
-		}
-	}
 }
